feat: 상품 관련 커스텀 예외 추가
상품 삭제 시 해당 상품이 없을 경우 에러 반환

+++ src/main/java/homeTry/product/service/AdminProductService.java
@@ -2,6 +2,7 @@ package homeTry.product.service;
 
 import homeTry.product.dto.request.ProductRequest;
 import homeTry.product.dto.response.ProductResponse;
+import homeTry.product.exception.badRequestException.ProductNotFoundException;
 import homeTry.product.model.entity.Product;
 import homeTry.product.repository.ProductRepository;
 import org.springframework.data.domain.Page;
@@ -31,6 +32,9 @@ public class AdminProductService {
 
     // 상품 삭제
     public void deleteProduct(Long productId) {
+        if (!productRepository.existsById(productId)) {
+            throw new ProductNotFoundException();
+        }
         productRepository.deleteById(productId);
     }
